Finish refactoring Math/Trig tests into separate files
