+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/MInverseTest.php
